refactor(device-settings): Extract setting controls into own files
